+++ src/rexecop/storage/__init__.py
@@ -1,3 +1,13 @@
+from rexecop.storage.factory import create_store, resolve_storage_backend
 from rexecop.storage.file_store import FileStore
+from rexecop.storage.port import OperationStoragePort, RuntimeStore
+from rexecop.storage.sqlite_store import SqliteStore
 
-__all__ = ["FileStore"]
+__all__ = [
+    "FileStore",
+    "OperationStoragePort",
+    "RuntimeStore",
+    "SqliteStore",
+    "create_store",
+    "resolve_storage_backend",
+]